Use long long for count

+++ src/person.cpp
@@ -19,8 +19,10 @@ double calculate_average_eligible_salary_ranges(
   const double total_eligible_salary = std::accumulate(
       eligible_salaries_view.begin(), eligible_salaries_view.end(), 0.0);
 
-  // To get the count, we count the view elements directly
-  const long count = std::ranges::distance(eligible_salaries_view);
+  // To get the count, we count the view elements directly. For more about
+  // the double-long usage, please refer to this SO post:
+  // https://stackoverflow.com/a/62132391/2748860
+  const long long count = std::ranges::distance(eligible_salaries_view);
 
   if (count == 0) {
     return 0.0;
